refactor: modify initial architecture
