+++ .env.example
@@ -1,3 +1,2 @@
-# opções para o rate-limit
-RATE_LIMIT_MAX= # 500
-RATE_LIMIT_INTERVAL= # 3600000 (1 hora)
+# url do backend
+BACKEND_URL= # http://localhost:5432
