+++ package.json
@@ -20,7 +20,6 @@
 		"axios": "^1.6.5",
 		"class-variance-authority": "^0.7.0",
 		"clsx": "^2.1.0",
-		"lru-cache": "^10.1.0",
 		"lucide-react": "^0.311.0",
 		"next": "14.0.4",
 		"next-themes": "^0.2.1",
@@ -33,7 +32,7 @@
 	"devDependencies": {
 		"@ianvs/prettier-plugin-sort-imports": "^4.1.1",
 		"@total-typescript/ts-reset": "^0.5.1",
-		"@types/node": "^20.11.4",
+		"@types/node": "^20.11.5",
 		"@types/react": "^18.2.48",
 		"@types/react-dom": "^18.2.18",
 		"@typescript-eslint/eslint-plugin": "^6.19.0",
@@ -42,7 +41,7 @@
 		"eslint": "^8.56.0",
 		"eslint-config-next": "^14.0.4",
 		"postcss": "^8.4.33",
-		"prettier": "^3.2.2",
+		"prettier": "^3.2.4",
 		"prettier-plugin-tailwindcss": "^0.5.11",
 		"tailwindcss": "^3.4.1",
 		"typescript": "^5.3.3"
