+++ src/middleware.ts
@@ -1,9 +1,8 @@
+import { loggingMiddleware } from '@/features/logging/logging-middleware'
 import { createMiddleware } from '@/libs/middleware'
-import { loggingMiddleware } from '@/middlewares/logging-middleware'
-import { rateLimitMiddleware } from '@/middlewares/rate-limit-middleware'
 
 export default createMiddleware({
-	'*': [loggingMiddleware, rateLimitMiddleware]
+	'*': [loggingMiddleware]
 })
 
 export const config = {
